Updated Financial Builder to use long rather than int due to the size of some of our foundations/orgs. I also decided to comment out most of the payload requests as the API calls were causing issues. I renamed GuideStarCLientTest to functionalTest so we can test that actual API calls work there. Consequently, tests are much faster and no longer time out.

+++ Candid.GuideStarAPI/Src/Helper/SearchPayload.cs
@@ -96,7 +96,7 @@ namespace Candid.GuideStarAPI
 
   public class Min_Max
   {
-    public int? min { get; set; }
-    public int? max { get; set; }
+    public long? min { get; set; }
+    public long? max { get; set; }
   }
 }
